Finished 2nd exercise
Script stops running when pattern is not matched in the second parameter.
Also, the .tmp file is deleted when an EXIT, SIGQUIT, SIGKILL or SIGTERM signal is sent to the script.

Also, the awk command is the one to go for getting the 5th and the 9th column for the $size and $path parameters to correctly and consistently use.

+++ it22105_ex_2.sh
@@ -2,6 +2,12 @@
 
 declare -r tmp=".tmp"
 
+function del_tmp {
+    rm -rf $tmp
+}
+
+trap "del_tmp" EXIT SIGQUIT SIGKILL SIGTERM
+
 if [ $# -ne 2 ]
 then
     echo -e "Usage: $0 direcotry-name file-size"
@@ -21,16 +27,30 @@ fi
 # for find: -size n[c --> bytes, w --> two-byte words, k --> kilobytes, M --> megabytes, G --> gigabytes]
 
 # Checking if 2nd parameter is a size that exists
-# if ! [[  ]] 
+# pattern looks for atleast one digit and up, after that will look for
+pattern='[[:digit:]]+[bcwkMG]'
+if ! [[ $2 =~ $pattern ]] 
+then
+    echo -e "Error: file-size input was incorrect. Correct input is in this format: n[cwbkMG]\n
+    \t n\tfor units of space\n
+    \t'b'\tfor 512-byte blocks (default)\n
+    \t'c'\tfor bytes\n
+    \t'w'\tfor two-byte words\n
+    \t'k'\tfor kilobytes\n
+    \t'M'\tfor megabytes\n
+    \t'G'\tfor gigabytes"
+    exit
+fi
 
 find "$dir" -type f -size +"$file_size" -exec ls -lh {} + 2> /dev/null > $tmp
 
 while IFS= read -r line
 do
-    size=$(echo "$line" | cut -d " " -f 5)
-    printf '%s\n' "$size"
+    size=$(echo "$line" | awk '{print $5}')
+    printf '%s\n' "Size of file: $size"
 
-    path=$(echo "$line" | cut -d " " -f 10-)
+    path=$(echo "$line" | awk '{print $9}')
+    printf '%s\n' "Path of file: $path"
     
     # THe read in the while line is occupying the STDIN, so the rm -i command is never given the STDIN, that's why I enforce the command to address the terminal directly
     rm -i "$path" < /dev/tty
